Moved JSON deserialisation to utilities module.
